Restructure arrow-c-data for package conflicts
